backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gspread==6.0.0
google-auth==2.27.0
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
python-dotenv==1.0.0
